Readability for deaf granny program
We take care of Granny answering silence, but also take care of how the
running code reads.  We create methods to help express the ideas in the
running code, and when you read it, you understand the idea of the
programs steps.  If you want to see implementation, you can read the
methods, though we also pay attention to how the methods read as well.

We use rubocop for style, and disable some cops when appropriate.