Age now starts with an actual age
This removes the need to give the number of seconds for each method call
on Age class.

`.to_s` gives us the opinion of what we think should be repesented as a
string.

year_months gives the ability for the user to format their strings as
they see fit, adding flexibility and freedom.

Add some tests.  May still need to add more, but this should cover most
bugs.

Lesson learned:  An array can be used to return multiple values back to
the user, because Ruby will only return one value, not multiples.

Also, splat operator `*` will flatten an array within an array.