No If..Else and no ?: ternary
We can solve this using just a logical statement using `or` and `and` or
`||` and `&&` and this will evaluate the line using a "short circuit" or
a logic gate.

When doing flow control, it is often that we will use the lower
precedence `or` and `and` flow control operators, and when using the
conditional statements `||` and `&&` we will likely be using them for
`if` statements or assignment statements.

An example of this is needing something assigned, but needing to operate
on the condition of the code.  For example:

`if 'q' == answer = gets.chomp`
Then we have our `answer` variable assigned that we can use inside of
your `if` statement, and don't need to assign outside of that statement.

Substitute `if` with `while` or `until` and you start to see the
benefit.