Working on Spec change
Added methods to override the shape class in Amoeba class.

Subject Line: Limit to 50 chars

Description, please reference the Issue or feature, along with detailed
information of the idea of the commit, remember, you may need to read this
later to get re-informed of this commit.

So to give an idea of what a good commit message looks like, let's pretend the
following lines is a commit message:

--------------------- <Snip> --------------------------------------------
Gives an idea of a good commit message
                                                 < -- This line must be blank!
We need to make sure that the sumamry line (the first line) is only 50
characters long or less, and separated from this body (detailed message) by a
blank line.  The reason for this is that the GUI tools depend on having a well
formed summary line for their purposes.

This detailed body should talk about the details of the idea, not really the
details of the code itself, as the code detail can be gotten by reading the
code.  We want to send ourselves a great message for when we come back and read
about why this commit was done, and what the intent of the code is.

As you can see from this commit message, we have a good summary of the idea of
this commit, simply that this commit gives an idea of a good commit message.

And of course, this is a "meta" commit, so the "code" really is this commit
message itself.  And the details of the idea of this commit are nicely detailed
here.
--------------------- <Snip> --------------------------------------------

Feel free (Please!!) to delete these lines, and avoid committing this message
as it is.  This is provided as a template to remind you to write good commit
messages!

And now, for your regularly scheduled commit information...