Removed the 'reporting' aspect of the method
The method extractions in the original implementation leads to me being
able to pass in a file, and being explicit about that.  But really, it
will accept anything that is 'file like', it simply needs to respond to
`mtime` method.

To demonstrate, I require 'pathname' and then use the projects parent
folder to determine the last modified date.