Modified Text Analyzer to add logging feature
Make use of Logger class to add logging feature.
New method of Logger class creates an instance of the class. Weneed to
pass the filename and number of old files to keep. It may be daily, weekly or
monthly. The info method of Logger class is used to log an info message.